requirements.txt
requests>=2.25.1
typing-extensions>=3.7.4
python-dotenv>=0.19.0